+++ ihp-typed-sql/default.nix
@@ -3,7 +3,7 @@
 , hasql-mapping, hasql-pool, hspec, ihp, fast-logger, lib
 , postgresql-libpq, postgresql-syntax, postgresql-types, process
 , scientific, string-conversions, template-haskell, temporary
-, temporary-ospath, text
+, temporary-ospath, text, wai
 }:
 mkDerivation {
   pname = "ihp-typed-sql";
@@ -13,7 +13,7 @@ mkDerivation {
     aeson base bytestring containers directory filepath haskell-src-meta hasql
     hasql-dynamic-statements hasql-mapping hasql-pool ihp
     postgresql-libpq postgresql-syntax postgresql-types process scientific
-    string-conversions template-haskell temporary text
+    string-conversions template-haskell temporary text wai
   ];
   testHaskellDepends = [
     base containers directory filepath hspec ihp fast-logger process